+++ packages/terra-core-docs/src/terra-dev-site/doc/badge/example/BadgeVisuallyHiddenText.jsx
@@ -1,15 +1,24 @@
 import React from 'react';
 import Badge from 'terra-badge';
+import classNames from 'classnames/bind';
+import styles from './Badge.module.scss';
+
+const cx = classNames.bind(styles);
 
 const BadgeVisuallyHiddenText = () => (
-  <div>
-    <Badge text="Low" visuallyHiddenText="Risk Factor Low" />
-    {' '}
-    <Badge text="Medium" visuallyHiddenText="Risk Factor Medium" />
-    {' '}
-    <Badge text="High" visuallyHiddenText="Risk Factor High" />
-    {' '}
-    <Badge text="Critical" visuallyHiddenText="Risk Factor Critical" />
+  <div className={cx('badge-container')}>
+    <div className={cx('badge')}>
+      <Badge text="Low" visuallyHiddenText="Risk Factor Low" />
+    </div>
+    <div className={cx('badge')}>
+      <Badge text="Medium" visuallyHiddenText="Risk Factor Medium" />
+    </div>
+    <div className={cx('badge')}>
+      <Badge text="High" visuallyHiddenText="Risk Factor High" />
+    </div>
+    <div className={cx('badge')}>
+      <Badge text="Critical" visuallyHiddenText="Risk Factor Critical" />
+    </div>
   </div>
 );